refactor(#42): SessionKey, CurrentUserId 파일 이동
CurrentArgumentResolver
1. CurrentUserIdArgumentResolver 는 CurrentUserId Annotation이 붙은 인자를 리졸빙하기 위한 커스텀 ArgumentResolver 입니다.
2. CurrentUserIdArgumentResolver 를 User Package 를 넣으려고 했으나, 해당 ArgumentResolver는 "User" Domain에 특정되어 있는 기술이 아니고, "인증 기술"에 관한 기술이 구현되어 있습니다.
3. 즉, Session을 사용하든 JWT를 사용하든 사용한 기술을 통해서 저장되어 있는 유저 정보를 가져오는 ArgumentResolver이라 User Domain에 국한되어 있지 않고, Web에 국한되어 있다고 판단하여 Common으로 옮겼습니다.
   - 또한, 해당 ArgumentResolver는 여러 도메인에서 공통적으로 자주 사용할 가능성이 높은 것도 옮긴 이유가 됩니다.
CurrentUserId
1. CurrentArgumentResolver가 Common Package에 있음에 따라, 이와 한 쌍인 CurrentUserId Annotation도 Common에 위치해두었습니다.
SessionKey
1. Session은 "User" Domain에 특정되어 있는 기술이 아니라, "인증 관련"에 특정되어 있는 기술입니다.
2. "웹", "인증 관련"에 특화되어 있는 기술이라 Common Package 으로 옮겼습니다.
