+++ src/main/java/com/airjnc/common/util/constant/SessionKey.java
@@ -1,0 +1,5 @@
+package com.airjnc.common.util.constant;
+
+public enum SessionKey {
+    USER,
+}
