--- src/main/java/com/airjnc/user/util/SessionKey.java
@@ -1,5 +1,0 @@
-package com.airjnc.user.util;
-
-public enum SessionKey {
-    USER,
-}
